Remove SignedCookiesMiddleware.
The only cookies we do set (csrftoken and sessionid) are exepmt from
signing, so this middleware is just hanging around doing nothing.

Django 1.4 introduced signed cookies, which can be used without
middleware. The only thing which nees to be done to use a signed cookie
is to replace the function pair get_cookie()/set_cookie() with
get_signed_cookie()/set_signed_cookie().