Fix file upload.
In python3 the name property of a TemporaryFile is an integer (the file
descriptor number), while in python2 the name was None. The UploadedFile
class tries to find out the basename of a file via os.basename, when it
is not None, so it fails horribly when the name is not a string.