SECURITY FIX: Enforce document visibility restrictions.
The current code relies on ecs.auth_conf to restrict the accessible
documents for every user, but the DocumentQFactory moronically returns
all the documents, resulting in the sad situation, that every user can
download every single document there is in the system. They just have to
guess the correct url.

This diff avoids ecs.auth_conf for simplicity reasons; the rules are
just too complex, for getting the QFactory correct. Instead, for every
source of documents, there is a separate route for download, which
should make it easier to reason about the different cases.