Get rid of the assigned tasks category.
Assigned tasks are now shown in 'my tasks', but marked with an icon to
reject it.

This prevents an edge case where a board member is marked as biased for
a submission for which he/she has a task assigned. In that case the task
would be in limbo - it can't be perfomed and not even delegated to
another user.