Import django-widget-tweaks library.
This allows to add the form-control class to input fields in a django
form, e.g. a pattern like this:

	{% for field in form %}
	    ...
	    {{ field }}
	    ...
	{% endfor %}

can be replaced with the following to produce a form with bootstrap
styled elements:

	{% load widget_tweaks %}
	...
	{% for field in form %}
	    ...
	    {{ field|add_class:'form-control' }}
	    ...
	{% endfor %}

Wherever we use a plain {{ form }}, it has to be unrolled by iterating
over the fields (as shown above). This is somewhat inconvenient and
verbose, however we only have six occurrences of a plain {{ form }} in
the tree.