Rework generation of communication messages.
Remove the {% strip %} and {% dedent %} template tags. They do not work
as expected and are a bad idea in general.

While there simplify the message templates by not using i18n. Wording
can be changed there. If we ever want to translate them the django
translation mechanism is the wrong place to do that.