Go full paranoia with our weasyprint url fetcher.
Stop using default_url_fetcher() to disallow the usage of http://,
file:// and anything else that would allow an attacker to get access to
files we don't explicitly export. The only allowed URI schemes are data:
and static:.

Also: Raise ValueError instead of IOError to satisfy the url fetcher
contract.