Remove the special external review button.
This is not necessary anymore because external review tasks are now
displayed in the task widget like any other task. Reviewers can still
start the review by clicking on the link in the message or via the task
widget.

With the header actions gone, we can now display the header title over
the full width. 100%!